requirements.txt
customtkinter
pillow
mysql-connector-python
